mimex-pixmc/requirements.txt
gym
hydra-core>=1.1
iopath
imageio-ffmpeg
omegaconf
setuptools
tensorboard
timm
